Completed transfaring the builder helper functions from apio_env.py to plugin_util.py
