+++ apio/scons/apio_env.py
@@ -17,18 +17,9 @@
 # pylint: disable=W0613
 
 import os
-import re
-import sys
-import json
-from typing import Dict, Tuple, List, Optional
-from dataclasses import dataclass
-from click import secho, style
-from SCons.Node import NodeList
-from SCons.Node.FS import File
-from SCons.Node.Alias import Alias
+from typing import Dict, List
+from click import secho
 from SCons.Script.SConscript import SConsEnvironment
-from SCons.Action import FunctionAction, Action
-from SCons.Builder import Builder
 from SCons.Environment import BuilderWrapper
 import SCons.Defaults
 from apio.scons.apio_args import ApioArgs
@@ -43,15 +34,6 @@ BUILD_DIR_SEP = BUILD_DIR + os.sep
 
 # -- Target name. This is the base file name for various build artifacts.
 TARGET = BUILD_DIR_SEP + "hardware"
-
-
-@dataclass(frozen=True)
-class SimulationConfig:
-    """Simulation parameters, for sim and test commands."""
-
-    testbench_name: str  # The testbench name, without the 'v' suffix.
-    build_testbench_name: str  # testbench_name prefixed by build dir.
-    srcs: List[str]  # List of source files to compile.
 
 
 # pylint: disable=too-many-public-methods
@@ -129,52 +111,6 @@ class ApioEnv:
         """Adds a dependency of one target on another."""
         self.env.Depends(target, dependency)
 
-    def map_params(self, params: Optional[List[str]], fmt: str) -> str:
-        """A common function construct a command string snippet from a list
-        of arguments. The functon does the following:
-        1. If params arg is None replace it with []
-        2. Drops empty or white space only items.
-        3. Maps the items using the format string which contains exactly one
-            placeholder {}.
-        4. Joins the items with a white space char.
-
-        For examples, see the unit test at test_scons_util.py.
-        """
-        # None designates empty list. Avoiding the pylint non safe default
-        # warning.
-        if params is None:
-            params = []
-
-        # Drop empty params and map the rest using the format string.
-        mapped_params = [fmt.format(x.strip()) for x in params if x.strip()]
-
-        # Join using a single space.
-        return " ".join(mapped_params)
-
-    def basename(self, file_name: str) -> str:
-        """Given a file name, returns it with the extension removed."""
-        result, _ = os.path.splitext(file_name)
-        return result
-
-    def is_verilog_src(
-        self, file_name: str, *, include_sv: bool = True
-    ) -> bool:
-        """Given a file name, determine by its extension if it's a verilog
-        source file (testbenches included).  If include_sv is True, include
-        also system verilog files."""
-        _, ext = os.path.splitext(file_name)
-        if include_sv:
-            return ext in [".v", ".sv"]
-        return ext in [".v"]
-
-    def has_testbench_name(self, file_name: str) -> bool:
-        """Given a file name, return true if it's base name indicates a
-        testbench. It can be for example abc_tb.v or _build/abc_tb.out.
-        The file extension is ignored.
-        """
-        name, _ = os.path.splitext(file_name)
-        return name.lower().endswith("_tb")
-
     def dump_env_vars(self) -> None:
         """Prints a list of the environment variables. For debugging."""
         dictionary = self.env.Dictionary()
@@ -185,422 +121,3 @@ class ApioEnv:
         for key in keys:
             print(f"{key} = {self.env[key]}")
         secho("<<< Env vars END\n", fg="magenta", color=True)
-
-    def programmer_cmd(self) -> str:
-        """Return the programmer command as derived from the scons "prog"
-        arg."""
-
-        # Get the programer command template arg.
-        prog_arg = self.args.PROG
-
-        # If empty then return as is. This must be an apio command that
-        # doesn't use the programmer.
-        if not prog_arg:
-            return prog_arg
-
-        # It's an error if the programmer command doesn't have the $SOURCE
-        # placeholder when scons inserts the binary file name.
-        if "$SOURCE" not in prog_arg:
-            secho(
-                "Error: [Internal] 'prog' argument does not contain "
-                f"the '$SOURCE' marker. [{prog_arg}]",
-                fg="red",
-                color=True,
-            )
-            sys.exit(1)
-
-        return prog_arg
-
-    def source_file_issue_action(self) -> FunctionAction:
-        """Returns a SCons action that scans the source files and print
-        error or warning messages about issues it finds."""
-
-        # A regex to identify "$dumpfile(" in testbenches.
-        testbench_dumpfile_re = re.compile(r"[$]dumpfile\s*[(]")
-
-        def report_source_files_issues(
-            source: List[File],
-            target: List[Alias],
-            env: SConsEnvironment,
-        ):
-            """The scanner function.."""
-
-            for file in source:
-
-                # -- For now we report issues only in testbenches so skip
-                # -- otherwise.
-                if not self.is_verilog_src(
-                    file.name
-                ) or not self.has_testbench_name(file.name):
-                    continue
-
-                # -- Read the testbench file text.
-                file_text = file.get_text_contents()
-
-                # -- if contains $dumpfile, print a warning.
-                if testbench_dumpfile_re.findall(file_text):
-                    secho(
-                        f"Warning: [{file.name}] Using $dumpfile() in apio "
-                        "testbenches is not recomanded.",
-                        fg="magenta",
-                        color=True,
-                    )
-
-        return Action(report_source_files_issues, "Scanning for issues.")
-
-    def vlt_path(self, path: str) -> str:
-        """Normalize a path that is used in the verilator config file
-        hardware.vlt. On windows it replaces "\" with "/". Otherwise it
-        leaves the path as is.
-        """
-        return path.replace("\\", "/")
-
-    def make_verilator_config_builder(self, config_text: str):
-        """Create a scons Builder that writes a verilator config file
-        (hardware.vlt) with the given text."""
-
-        def verilator_config_func(target, source, env):
-            """Creates a verilator .vlt config files."""
-            with open(
-                target[0].get_path(), "w", encoding="utf-8"
-            ) as target_file:
-                target_file.write(config_text)
-            return 0
-
-        return Builder(
-            action=Action(
-                verilator_config_func, "Creating verilator config file."
-            ),
-            suffix=".vlt",
-        )
-
-    def source_files(self) -> Tuple[List[str], List[str]]:
-        """Get the list of *.v files, splitted into synth and testbench lists.
-        If a .v file has the suffix _tb.v it's is classified st a testbench,
-        otherwise as a synthesis file.
-        """
-        # -- Get a list of all *.v and .sv files in the project dir.
-        files: List[File] = self.env.Glob("*.sv")
-        if files:
-            secho(
-                "Warning: project contains .sv files, system-verilog support "
-                "is experimental.",
-                fg="yellow",
-                color=True,
-            )
-        files = files + self.env.Glob("*.v")
-
-        # Split file names to synth files and testbench file lists
-        synth_srcs = []
-        test_srcs = []
-        for file in files:
-            if self.has_testbench_name(file.name):
-                test_srcs.append(file.name)
-            else:
-                synth_srcs.append(file.name)
-        return (synth_srcs, test_srcs)
-
-    def get_sim_config(
-        self, testbench: str, synth_srcs: List[str]
-    ) -> SimulationConfig:
-        """Returns a SimulationConfig for a sim command. 'testbench' is
-        a required testbench file name. 'synth_srcs' is the list of all
-        module sources as returned by get_source_files()."""
-        assert testbench, "Missing testbench name"
-
-        # Construct a SimulationParams with all the synth files + the
-        # testbench file.
-        testbench_name = self.basename(testbench)
-        build_testbench_name = BUILD_DIR_SEP + testbench_name
-        srcs = synth_srcs + [testbench]
-        return SimulationConfig(testbench_name, build_testbench_name, srcs)
-
-    def get_tests_configs(
-        self,
-        testbench: str,
-        synth_srcs: List[str],
-        test_srcs: list[str],
-    ) -> List[SimulationConfig]:
-        """Return a list of SimulationConfigs for each of the testbenches that
-        need to be run for a 'apio test' command. If testbench is empty,
-        all the testbenches in test_srcs will be tested. Otherwise, only the
-        testbench in testbench will be tested. synth_srcs and test_srcs are
-        source and test file lists as returned by get_source_files()."""
-        # List of testbenches to be tested.
-        if testbench:
-            testbenches = [testbench]
-        else:
-            testbenches = test_srcs
-
-        # -- If this fails, it's a programming error.
-        assert testbenches, "No testbenches"
-
-        # Construct a config for each testbench.
-        configs = []
-        for tb in testbenches:
-            testbench_name = self.basename(tb)
-            build_testbench_name = BUILD_DIR_SEP + testbench_name
-            srcs = synth_srcs + [tb]
-            configs.append(
-                SimulationConfig(testbench_name, build_testbench_name, srcs)
-            )
-
-        return configs
-
-    def waves_target(
-        self,
-        name: str,
-        vcd_file_target: NodeList,
-        sim_config: SimulationConfig,
-        allways_build: bool = False,
-    ) -> List[Alias]:
-        """Construct a target to launch the QTWave signal viwer.
-        vcd_file_target is the simulator target that generated the vcd file
-        with the signals. Returns the new targets.
-        """
-
-        # -- Construct the commands list.
-        commands = []
-
-        # -- On windows we need to setup the cache. This could be done once
-        # -- when the oss-cad-suite is installed but since we currently don't
-        # -- have a package setup mechanism, we do it here on each invocation.
-        # -- The time penalty is negligible.
-        # -- With the stock oss-cad-suite windows package, this is done in the
-        # -- environment.bat script.
-        if self.is_windows:
-            commands.append("gdk-pixbuf-query-loaders --update-cache")
-
-        # -- The actual wave viewer command.
-        commands.append(
-            "gtkwave {0} {1} {2}.gtkw".format(
-                '--rcvar "splash_disable on" --rcvar "do_initial_zoom_fit 1"',
-                vcd_file_target[0],
-                sim_config.testbench_name,
-            )
-        )
-
-        target = self.alias(
-            name,
-            source=vcd_file_target,
-            action=commands,
-            allways_build=allways_build,
-        )
-
-        return target
-
-    # pylint: disable=too-many-arguments
-    # pylint: disable=too-many-positional-arguments
-    def iverilog_action(
-        self,
-        *,
-        verbose: bool,
-        vcd_output_name: str,
-        is_interactive: bool,
-        extra_params: List[str] = None,
-        lib_dirs: List[str] = None,
-        lib_files: List[str] = None,
-    ) -> str:
-        """Construct an iverilog scons action string.
-        * env: Rhe scons environment.
-        * verbose: IVerilog will show extra info.
-        * vcd_output_name: Value for the macro VCD_OUTPUT.
-        * is_interactive: True for apio sim, False otherwise.
-        * extra_params: Optional list of additional IVerilog params.
-        * lib_dirs: Optional list of dir pathes to include.
-        * lib_files: Optional list of library files to compile.
-        *
-        * Returns the scons action string for the IVerilog command.
-        """
-
-        # Escaping for windows. '\' -> '\\'
-        escaped_vcd_output_name = vcd_output_name.replace("\\", "\\\\")
-
-        # -- Construct the action string.
-        # -- The -g2012 is for system-verilog support and we use it here as an
-        # -- experimental feature.
-        action = (
-            "iverilog -g2012 {0} -o $TARGET {1} {2} {3} {4} {5} $SOURCES"
-        ).format(
-            "-v" if verbose else "",
-            f"-DVCD_OUTPUT={escaped_vcd_output_name}",
-            "-DINTERACTIVE_SIM" if is_interactive else "",
-            self.map_params(extra_params, "{}"),
-            self.map_params(lib_dirs, '-I"{}"'),
-            self.map_params(lib_files, '"{}"'),
-        )
-
-        return action
-
-    # pylint: disable=too-many-arguments
-    # pylint: disable=too-many-positional-arguments
-    def verilator_lint_action(
-        self,
-        *,
-        warnings_all: bool = False,
-        warnings_no_style: bool = False,
-        no_warns: List[str] = None,
-        warns: List[str] = None,
-        top_module: str = "",
-        extra_params: List[str] = None,
-        lib_dirs: List[str] = None,
-        lib_files: List[str] = None,
-    ) -> str:
-        """Construct an verilator scons action string.
-        * env: Rhe scons environment.
-        * warnings_all: If True, use -Wall.
-        * warnings_no_style: If True, use -Wno-style.
-        * no_warns: Optional list with verilator warning codes to disble.
-        * warns: Optional list with verilator warning codes to enable.
-        * top_module: If not empty, use --top-module <top_module>.
-        * extra_params: Optional additional arguments.
-        * libs_dirs: Optional directories for include search.
-        * lib_files: Optional additional files to include.
-        """
-
-        action = (
-            "verilator_bin --lint-only --quiet --bbox-unsup --timing "
-            "-Wno-TIMESCALEMOD -Wno-MULTITOP "
-            "{0} {1} {2} {3} {4} {5} {6} {7} {8} $SOURCES"
-        ).format(
-            "-Wall" if warnings_all else "",
-            "-Wno-style" if warnings_no_style else "",
-            self.map_params(no_warns, "-Wno-{}"),
-            self.map_params(warns, "-Wwarn-{}"),
-            f"--top-module {top_module}" if top_module else "",
-            self.map_params(extra_params, "{}"),
-            self.map_params(lib_dirs, '-I"{}"'),
-            TARGET + ".vlt",
-            self.map_params(lib_files, '"{}"'),
-        )
-
-        return [self.source_file_issue_action(), action]
-
-    # pylint: disable=too-many-locals
-    def _print_pnr_report(
-        self,
-        json_txt: str,
-        clk_name_index: int,
-        verbose: bool,
-    ) -> None:
-        """Accepts the text of the pnr json report and prints it in
-        a user friendly way. Used by the 'apio report' command."""
-        # -- Json text to tree of Dicts.
-        report: Dict[str, any] = json.loads(json_txt)
-
-        # --- Report utilization
-        secho("")
-        secho("UTILIZATION:", fg="cyan", color=True)
-        utilization = report["utilization"]
-        for resource, vals in utilization.items():
-            available = vals["available"]
-            used = vals["used"]
-            percents = int(100 * used / available)
-            fg = "magenta" if used > 0 else None
-            secho(
-                f"{resource:>20}: {used:5} {available:5} {percents:5}%",
-                fg=fg,
-                color=True,
-            )
-
-        # -- Report max clock speeds.
-        # --
-        # -- NOTE: As of Oct 2024, some projects do not generate timing
-        # -- information and this is being investigated.
-        # -- See https://github.com/FPGAwars/icestudio/issues/774 for details.
-        secho("")
-        secho("CLOCKS:", fg="cyan", color=True)
-        clocks = report["fmax"]
-        if len(clocks) > 0:
-            for clk_net, vals in clocks.items():
-                # -- Extract clock name from the net name.
-                clk_signal = clk_net.split("$")[clk_name_index]
-
-                # -- Report speed.
-                max_mhz = vals["achieved"]
-                styled_max_mhz = style(f"{max_mhz:7.2f}", fg="magenta")
-                secho(f"{clk_signal:>20}: {styled_max_mhz} Mhz max")
-
-        # -- For now we ignore the critical path report in the pnr report and
-        # -- refer the user to the pnr verbose output.
-        secho("")
-        if not verbose:
-            secho(
-                "Use 'apio report --verbose' for more details.",
-                fg="yellow",
-                color=True,
-            )
-
-    def report_action(
-        self, clk_name_index: int, verbose: bool
-    ) -> FunctionAction:
-        """Returns a SCons action to format and print the PNR reort from the
-        PNR json report file. Used by the 'apio report' command.
-        'script_id' identifies the calling SConstruct script and 'verbose'
-        indicates if the --verbose flag was invoked."""
-
-        def print_pnr_report(
-            source: List[File],
-            target: List[Alias],
-            env: SConsEnvironment,
-        ):
-            """Action function. Loads the pnr json report and print in a user
-            friendly way."""
-            unused(target, env)
-            json_file: File = source[0]
-            json_txt: str = json_file.get_text_contents()
-            self._print_pnr_report(json_txt, clk_name_index, verbose)
-
-        return Action(print_pnr_report, "Formatting pnr report.")
-
-    def clean_if_requested(self) -> None:
-        """Should be called only when the "clean" target is specified.
-        Configures in scons env do delete all the files in the build directory.
-        """
-
-        # -- We perform cleaning only when the scons 'clean' option is active.
-        if not self.env.GetOption("clean"):
-            return
-
-        # -- Get the list of all files to clean. Scons adds to the list non
-        # -- existing files from other targets it encountered.
-        files_to_clean = (
-            self.env.Glob(f"{BUILD_DIR_SEP}*")
-            + self.env.Glob("zadig.ini")
-            + self.env.Glob(".sconsign.dblite")
-            + self.env.Glob("_build")
-        )
-
-        # pylint: disable=fixme
-        # -- TODO: Remove the cleanup of legacy files after releasing the first
-        # -- release with the _build directory.
-        # --
-        # --
-        # -- Until apio 0.9.6, the build artifacts were created in the project
-        # -- directory rather than the _build directory. To simplify the
-        # -- transition we clean here also left over files from 0.9.5.
-        legacy_files_to_clean = (
-            self.env.Glob("hardware.*")
-            + self.env.Glob("*_tb.vcd")
-            + self.env.Glob("*_tb.out")
-        )
-
-        if legacy_files_to_clean:
-            secho(
-                "Deleting also leftover files.",
-                fg="yellow",
-                color=True,
-            )
-
-            files_to_clean.extend(legacy_files_to_clean)
-
-        # -- Create a dummy target.  I
-        dummy_target = self.env.Command("cleanup-target", "", "")
-
-        # -- Associate all the files with the dummy target.
-        self.env.Clean(dummy_target, files_to_clean)
-
-
-def unused(*_):
-    """Fake a use of an unused variable or argument."""
-    # pass
