+++ test/scons/test_apio_env.py
@@ -3,19 +3,12 @@ Tests of the scons ApioEnv.
 """
 
 import os
-from os.path import isfile, exists
-from pathlib import Path
 from typing import Dict
-from test.conftest import ApioRunner
-import pytest
 import SCons.Script.SConsOptions
-from SCons.Node.FS import FS
 import SCons.Node.FS
 import SCons.Environment
 import SCons.Defaults
 import SCons.Script.Main
-from SCons.Script import SetOption
-from pytest import LogCaptureFixture
 from apio.scons.apio_env import ApioEnv
 
 # pylint: disable=fixme
@@ -91,62 +84,6 @@ def make_test_apio_env(
     )
 
 
-def test_has_testbench_name():
-    """Tests the scons_util.test_is_testbench() method"""
-
-    env = make_test_apio_env()
-
-    # -- Testbench names
-    assert env.has_testbench_name("aaa_tb.v")
-    assert env.has_testbench_name("aaa_tb.out")
-    assert env.has_testbench_name("bbb/aaa_tb.v")
-    assert env.has_testbench_name("bbb\\aaa_tb.v")
-    assert env.has_testbench_name("aaa__tb.v")
-    assert env.has_testbench_name("Aaa__Tb.v")
-    assert env.has_testbench_name("bbb/aaa_tb.v")
-    assert env.has_testbench_name("bbb\\aaa_tb.v")
-
-    # -- Non testbench names.
-    assert not env.has_testbench_name("aaatb.v")
-    assert not env.has_testbench_name("aaa.v")
-
-
-def test_is_verilog_src():
-    """Tests the scons_util.is_verilog_src() method"""
-
-    e = make_test_apio_env()
-
-    # -- Verilog and system-verilog source names, system-verilog included.
-    assert e.is_verilog_src("aaa.v")
-    assert e.is_verilog_src("bbb/aaa.v")
-    assert e.is_verilog_src("bbb\\aaa.v")
-    assert e.is_verilog_src("aaatb.v")
-    assert e.is_verilog_src("aaa_tb.v")
-    assert e.is_verilog_src("aaa.sv")
-    assert e.is_verilog_src("bbb\\aaa.sv")
-    assert e.is_verilog_src("aaa_tb.sv")
-
-    # -- Verilog and system-verilog source names, system-verilog excluded.
-    assert e.is_verilog_src("aaa.v", include_sv=False)
-    assert e.is_verilog_src("bbb/aaa.v", include_sv=False)
-    assert e.is_verilog_src("bbb\\aaa.v", include_sv=False)
-    assert e.is_verilog_src("aaatb.v", include_sv=False)
-    assert e.is_verilog_src("aaa_tb.v", include_sv=False)
-    assert not e.is_verilog_src("aaa.sv", include_sv=False)
-    assert not e.is_verilog_src("bbb\\aaa.sv", include_sv=False)
-    assert not e.is_verilog_src("aaa_tb.sv", include_sv=False)
-
-    # -- Non verilog source names, system-verilog included.
-    assert not e.is_verilog_src("aaatb.vv")
-    assert not e.is_verilog_src("aaatb.V")
-    assert not e.is_verilog_src("aaa_tb.vh")
-
-    # -- Non verilog source names, system-verilog excluded.
-    assert not e.is_verilog_src("aaatb.vv", include_sv=False)
-    assert not e.is_verilog_src("aaatb.V", include_sv=False)
-    assert not e.is_verilog_src("aaa_tb.vh", include_sv=False)
-
-
 def test_env_args():
     """Tests the scons env args retrieval."""
 
@@ -182,66 +119,6 @@ def test_env_platform_id():
     assert env.is_windows
 
 
-def test_clean_if_requested(apio_runner: ApioRunner):
-    """Tests the success path of set_up_cleanup()."""
-
-    with apio_runner.in_sandbox():
-
-        # -- Create an env with 'clean' option set.
-        apio_env = make_test_apio_env()
-
-        # -- Create files that shouldn't be cleaned up.
-        Path("my_source.v")
-        Path("apio.ini")
-
-        # -- Create files that should be cleaned up.
-        Path("zadig.ini").touch()
-        Path("_build").mkdir()
-        Path("_build/aaa").touch()
-        Path("_build/bbb").touch()
-
-        # -- Run clean_if_requested with no cleanup requested. It should
-        # -- not add any target.
-        assert len(SconsHacks.get_targets()) == 0
-        apio_env.clean_if_requested()
-        assert len(SconsHacks.get_targets()) == 0
-
-        # -- Run the cleanup setup. It's expected to add a single
-        # -- target with the dependencies to clean up.
-        assert len(SconsHacks.get_targets()) == 0
-        SetOption("clean", True)
-        apio_env.clean_if_requested()
-        assert len(SconsHacks.get_targets()) == 1
-
-        # -- Get the target and its dependencies
-        items_list = list(SconsHacks.get_targets().items())
-        target, dependencies = items_list[0]
-
-        # -- Verify the tartget name, hard coded in set_up_cleanup()
-        assert target.name == "cleanup-target"
-
-        # -- Verif the dependencies. These are the files to delete.
-        file_names = [x.name for x in dependencies]
-        assert file_names == ["aaa", "bbb", "zadig.ini", "_build"]
-
-
-def test_map_params():
-    """Test the map_params() method."""
-
-    apio_env = make_test_apio_env()
-
-    # -- Empty cases
-    assert apio_env.map_params([], "x_{}_y") == ""
-    assert apio_env.map_params(["", "   "], "x_{}_y") == ""
-
-    # -- Non empty cases
-    assert apio_env.map_params(["a"], "x_{}_y") == "x_a_y"
-    assert apio_env.map_params([" a "], "x_{}_y") == "x_a_y"
-    assert (
-        apio_env.map_params(["a", "a", "b"], "x_{}_y") == "x_a_y x_a_y x_b_y"
-    )
-
-
 def test_targeting():
     """Test the targeting() method."""
 
@@ -250,95 +127,3 @@ def test_targeting():
 
     assert apio_env.targeting("build")
     assert not apio_env.targeting("upload")
-
-
-def test_get_programmer_cmd(capsys: LogCaptureFixture):
-    """Tests the method get_programmer_cmd()."""
-
-    # -- Without a "prog" arg, expected to return "". This is the case
-    # -- when scons handles a command that doesn't use the programmer.
-    apio_env = make_test_apio_env()
-    assert apio_env.programmer_cmd() == ""
-
-    # -- If prog is specified, expected to return it.
-    apio_env = make_test_apio_env(extra_args={"prog": "my_prog aa $SOURCE bb"})
-    assert apio_env.programmer_cmd() == "my_prog aa $SOURCE bb"
-
-    # -- If prog string doesn't contains $SOURCE, expected to exit with an
-    # -- error message.
-    apio_env = make_test_apio_env(extra_args={"prog": "my_prog aa SOURCE bb"})
-    with pytest.raises(SystemExit) as e:
-        capsys.readouterr()  # Reset capturing.
-        apio_env.programmer_cmd()
-    captured = capsys.readouterr()
-    assert e.value.code == 1
-    assert "does not contain the '$SOURCE'" in captured.out
-
-
-def test_make_verilator_config_builder(apio_runner: ApioRunner):
-    """Tests the make_verilator_config_builder() method."""
-
-    with apio_runner.in_sandbox() as sb:
-
-        # -- Create a test scons env.
-        apio_env = make_test_apio_env()
-
-        # -- Call the tested method to create a builder.
-        builder = apio_env.make_verilator_config_builder("test-text")
-
-        # -- Verify builder suffixes.
-        assert builder.suffix == ".vlt"
-        assert builder.src_suffix == []
-
-        # -- Create a target that doesn't exist yet.
-        assert not exists("hardware.vlt")
-        target = FS.File(FS(), "hardware.vlt")
-
-        # -- Invoke the builder's action to create the target.
-        builder.action(target, [], apio_env.env)
-        assert isfile("hardware.vlt")
-
-        # -- Verify that the file was created with the tiven text.
-        text = sb.read_file("hardware.vlt")
-
-        assert text == "test-text"
-
-
-def test_get_source_files(apio_runner):
-    """Tests the get_source_files() method."""
-
-    with apio_runner.in_sandbox():
-
-        # -- Create a test scons env.
-        apio_env = make_test_apio_env()
-
-        # -- Make files verilog src names (out of order)
-        Path("bbb.v").touch()
-        Path("aaa.v").touch()
-
-        # -- Make files with testbench names (out of order)
-        Path("ccc_tb.v").touch()
-        Path("aaa_tb.v").touch()
-
-        # -- Make files with non related names.
-        Path("ddd.vh").touch()
-        Path("eee.vlt").touch()
-        Path("subdir").mkdir()
-        Path("subdir/eee.v").touch()
-
-        # -- Invoked the tested method.
-        srcs, testbenches = apio_env.source_files()
-
-        # -- Verify results.
-        assert srcs == ["aaa.v", "bbb.v"]
-        assert testbenches == ["aaa_tb.v", "ccc_tb.v"]
-
-
-def test_vlt_path():
-    """Tests the vlt_path path string mapping."""
-
-    apio_env = make_test_apio_env()
-
-    assert apio_env.vlt_path("") == ""
-    assert apio_env.vlt_path("/aa/bb/cc.xyz") == "/aa/bb/cc.xyz"
-    assert apio_env.vlt_path("C:\\aa\\bb/cc.xyz") == "C:/aa/bb/cc.xyz"
